Andriod_Offloading/CloudServer/requirements.txt
Flask==2.2.5
Pillow==9.5.0
gunicorn==21.2.0
